requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn==0.34.0
python-dotenv==1.0.0
pydantic==2.5.3

# Multi-Agent Framework
crewai==0.11.2
langchain==0.1.9
langchain-anthropic==0.1.1
langchain-community==0.0.24

# LLM
anthropic==0.18.1

# Scraping
apify-client==1.6.4
httpx==0.26.0
beautifulsoup4==4.12.3
requests==2.31.0

# Database
pymongo==4.6.1
motor==3.3.2
pinecone-client==3.2.2

# NLP & Language Detection
langdetect==1.0.9
nltk==3.8.1
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Dashboard
streamlit==1.31.0
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.18.0

# Security & Auth
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
loguru==0.7.2
tenacity==8.2.3